--- src/test/individual-tests/query-test.ts
@@ -1,7 +1,0 @@
-import supertest = require('supertest');
-
-export const queryRequest = (query: string) => {
-  return supertest(`http://localhost:${process.env.PORT}`).post('/').send({
-    query,
-  });
-};
